feat: add util fns for acting and training plus support for chunkwise
