+++ mava/configs/network/ff_retention.yaml
@@ -5,11 +5,8 @@ n_head: 1 # Number of heads
 decay_scaling_factor: 0.8 # Decay scaling factor for the kappa parameter: kappa = kappa * decay_scaling_factor
 obs_prenorm_fn_name: rmsnorm # Observation pre-normalisation function name [layernorm, rmsnorm, identity]
 
-#TODO: update this once we merge scaling
-chunkwise:
-  # Chunking the sequence into smaller parts to fit into memory.
-  use_chunkwise: False # Whether to chunk the sequence.
-  chunk_size: 32 # Size of the chunk: calculated over timesteps dim.
-
 ff_config:
   type: "ff_sable" # Type of the network.
+  # Chunking the sequence into smaller parts to fit into memory.
+  use_chunkwise: False # Whether to chunk the sequence.
+  agents_chunk_size: 2 # Size of the chunk: calculated over timesteps dim.
